docs: add documentation for the LoadingState component
