+++ src/actions/Button.scss
@@ -274,12 +274,3 @@
     }
   }
 }
-
-@keyframes spinning-rotation {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
