+++ src/global/animations.scss
@@ -1,0 +1,8 @@
+@keyframes spinning-rotation {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
