+++ src/global/app-css.scss
@@ -90,3 +90,4 @@ body {
 
 @import "text.scss";
 @import "spacing.scss";
+@import "animations.scss";
