--- src/overlays/__stories__/LoadingState.docs.mdx
@@ -1,14 +1,0 @@
-import { Meta } from "@storybook/addon-docs"
-import { ArgTypes, Title, Subtitle, Description } from "@storybook/blocks"
-import LoadingState from "../LoadingState"
-import * as stories from "./LoadingState.stories.tsx"
-
-<Meta of={stories} />
-
-<Title />
-<Subtitle>
-  Loading stuff. Yeah!
-</Subtitle>
-Read more documentation about the LoadingState component in
-[Zeroheight](https://zeroheight.com/5e69dd4e1/p/86f988-dialog).
-
